[resources] :boom: Completely review font icons system
Aside from some renaming, the biggest change is that enums for icons are dropped in favor of properties files.

In my recent testings, it turned out that enums can be quite slow, especially if they are too big. FontAwesomeSolid has more than a thousand icons. When Java loads the enumerator, there's a noticeable overhead. I also learnt that every time the values() method is called, a copy of the original array is returned (to simulate immutability), which made things even worse.

MFXFontIcon was also slow to instantiate (slow here is relative though), so I slightly optimized its code too.

Properties are fast to load and to lookup since they are basically HashMaps. The only downside of this new system is that we lose the strong typing offered by enums. The only way to lookup and set icons now is via strings. But then again, this is no big deal, just go on the icon pack website, find the icon you need and copy its name (at least it's this easy for FontAwesome)

Another hot path was and still kinda is the registry for the various icon packs which allows MFXFontIcon to support any pack, even third party ones. When the class is first loaded, official packs are registered, and this operation can cause a bit of overhead. The documentation describes how you can improve this with a tradeoff

Signed-off-by: palexdev <alessandro.parisi406@gmail.com>