[components] :fire: Nuke virtualized menu content
It did not make much sense. Well, to be precise, the idea was to use the same menus infrastructure for controls such as combo boxes, split buttons, etc.
But that cannot be done because MFXMenuItem does not carry data, and that's crucial for choice-based controls

For standard menus, typically all items are visible, and if scrolling is needed, then it's enough to wrap the standard menu content in a scroll pane.

Signed-off-by: palexdev <alessandro.parisi406@gmail.com>